pyproject.toml
[tool.poetry]
name = "adeeny"
version = "0.1.0"
description = ""
authors = ["coderhanzo <reinerbielert@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
django-phonenumber-field = "^8.0.0"
djoser = "^2.2.3"
phonenumbers = "^8.13.40"
pillow = "^10.4.0"
djangorestframework-gis = "^1.0"
drf-extra-fields = "^3.7.0"
mysql-connector-python = "^9.0.0"
psycopg2 = "^2.9.9"
libmagic = "^1.0"
python-magic = "^0.4.27"
gunicorn = "^22.0.0"
whitenoise = "^6.7.0"
dj-database-url = "^2.2.0"
uvicorn = "^0.30.1"
django-cors-headers = "^4.4.0"
drf-base64 = "^2.0"
mysqlclient = "^2.2.4"
django-money = "^3.5.3"
python-decouple = "^3.8"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
